dom/tests/mochitest/bugs/test_bug304459.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=304459
-->
<head>
  <title>Test for Bug 304459</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=304459">Mozilla Bug 304459</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe id="frame"></iframe>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 304459 **/

function iframeLoaded()
{
  var frame = document.getElementById("frame");
  frame.onload = undefined;
  frame.src = "http://example.org/tests/dom/tests/mochitest/bugs/iframe_bug304459-2.html";
}

function receiveMessage(evt)
{
  ok(evt.data, "Prototype leaks across navigation");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

window.addEventListener("message", receiveMessage, false);

var frame = document.getElementById("frame");
frame.onload = iframeLoaded;
frame.src = "iframe_bug304459-1.html"

</script>
</pre>
</body>
</html>
